refactor(errors): extract isFileNotFoundError to shared utility
- Add isFileNotFoundError to @common/errors for reuse across codebase
- Simplify cleanupIndentLog by removing redundant displayPath parameter
- Remove local isFileNotFound in favor of shared utility
